Replace example link with real one
Carla intended, based on the information in
https://github.com/owncloud/doc-index-org/commit/b1320220, to not create
a real link, but to provide a shorthand guide to the links to the three
manuals. Based on https://github.com/owncloud/documentation/issues/3286,
I believe that, while well intended, it is causing confusion instead of
helping out. Given that, this commit changes the link to a working one.